+++ Mesh/MeshBuilder2.cs
@@ -1,8 +1,10 @@
-namespace Cutulu.Core
+namespace Cutulu.Mesh
 {
     using System.Collections.Generic;
     using System.Linq;
     using Godot;
+
+    using Core;
 
     /// <summary>
     /// Highly GPU optimized mesh generation tool for fast and easy use. Allows for fast polygon and wall meshes to be added.
