+++ UI/OptionPointer.cs
@@ -1,7 +1,9 @@
-using Godot;
-
-namespace Cutulu.Core.UX
+namespace Cutulu.UI
 {
+    using Godot;
+
+    using Core;
+
     [GlobalClass]
     public partial class OptionPointer : PolyControl
     {
